backend/requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.5.0
Markdown==3.3.4
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
zipp==3.4.1